Vis-work.py
adding the options in the run_and_plot function to select how many genes want to appear in the count per cell plot (ntop_genes). Choose to filter or not the genes in the box plot of ratio (filtred=True). Choose the size of the dots in the spatial plot, default is a good guess of the optimal dot size.